% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab4package-package.R
\docType{package}
\name{<lab4package>}
\alias{<lab4package>}
\title{lab4package}
\description{
This package has been built for the 4th lab assignment of the "Advanced Programming in R".
It contains the \code{linreg} class.
}
\details{
Written by: Arash Haratian & Daniel Díaz-Roncero González
}
